+++ bundle/demo.ts
@@ -1,0 +1,4 @@
+import {registerMediaWikiCore} from '../src/index';
+export {CodeMirror6} from '../src/index';
+
+registerMediaWikiCore();