+++ src/escape.ts
@@ -3,7 +3,8 @@ import {EditorSelection} from '@codemirror/state';
 import {indentMore, indentLess} from '@codemirror/commands';
 import {getLSP} from '@bhsd/browser';
 import {CodeMirror6} from './codemirror';
-import type {Command} from '@codemirror/view';
+import {menuRegistry} from './statusBar';
+import type {EditorView, Command} from '@codemirror/view';
 import type {Extension, SelectionRange} from '@codemirror/state';
 
 const entity = {'"': 'quot', "'": 'apos', '<': 'lt', '>': 'gt', '&': 'amp', ' ': 'nbsp'};
@@ -37,37 +38,85 @@ export const escapeHTML = (str: string): string => [...str].map(c => {
 		return encodeURIComponent(str);
 	};
 
+const escapeWiki = (cm: CodeMirror6): boolean => {
+	const view = cm.view!,
+		{state} = view,
+		{ranges} = state.selection,
+		lsp = getLSP(view, false, cm.getWikiConfig);
+	if (lsp && 'provideRefactoringAction' in lsp && ranges.some(({empty}) => !empty)) {
+		(async () => {
+			const replacements = new WeakMap<SelectionRange, string | undefined>();
+			for (const range of ranges) {
+				// eslint-disable-next-line no-await-in-loop
+				const [action] = await lsp.provideRefactoringAction(state.sliceDoc(range.from, range.to));
+				replacements.set(range, action?.edit!.changes!['']![0]!.newText);
+			}
+			view.dispatch(state.changeByRange(range => {
+				const insert = replacements.get(range);
+				if (insert === undefined) {
+					return {range};
+				}
+				return {
+					range: EditorSelection.range(range.from, range.from + insert.length),
+					changes: {from: range.from, to: range.to, insert},
+				};
+			}));
+		})();
+		return true;
+	}
+	return false;
+};
+
+const handlerBase = (view: EditorView, e: MouseEvent): void => {
+	e.stopPropagation();
+	view.focus();
+};
+
+let items: HTMLDivElement[] | undefined;
+
+menuRegistry.push({
+	name: 'escape',
+	isActionable({lang, view}): boolean {
+		return lang === 'mediawiki' && view!.state.selection.ranges.some(({empty}) => !empty);
+	},
+	getItems(cm): HTMLDivElement[] {
+		if (!items) {
+			const view = cm.view!,
+				btnHTML = document.createElement('div'),
+				btnURI = document.createElement('div');
+			btnHTML.textContent = 'HTML escape';
+			btnHTML.addEventListener('click', e => {
+				CodeMirror6.replaceSelections(view, escapeHTML);
+				handlerBase(view, e);
+			});
+			btnURI.textContent = 'URI encode/decode';
+			btnURI.addEventListener('click', e => {
+				CodeMirror6.replaceSelections(view, escapeURI);
+				handlerBase(view, e);
+			});
+			items = [btnHTML, btnURI];
+			const lsp = getLSP(view, false, cm.getWikiConfig);
+			if (lsp && 'provideRefactoringAction' in lsp) {
+				const btnWiki = document.createElement('div');
+				btnWiki.textContent = 'Escape with magic words';
+				btnWiki.addEventListener('click', e => {
+					escapeWiki(cm);
+					handlerBase(view, e);
+				});
+				items.unshift(btnWiki);
+			}
+		}
+		return items;
+	},
+});
+
 export default (cm: CodeMirror6): Extension => keymap.of([
 	{key: 'Mod-[', run: convert(escapeHTML, indentLess)},
 	{key: 'Mod-]', run: convert(escapeURI, indentMore)},
 	{
 		key: 'Mod-\\',
-		run(view): boolean {
-			const {state} = view,
-				{ranges} = state.selection,
-				lsp = getLSP(view, false, cm.getWikiConfig);
-			if (lsp && 'provideRefactoringAction' in lsp && ranges.some(({empty}) => !empty)) {
-				(async () => {
-					const replacements = new WeakMap<SelectionRange, string | undefined>();
-					for (const range of ranges) {
-						// eslint-disable-next-line no-await-in-loop
-						const [action] = await lsp.provideRefactoringAction(state.sliceDoc(range.from, range.to));
-						replacements.set(range, action?.edit!.changes!['']![0]!.newText);
-					}
-					view.dispatch(state.changeByRange(range => {
-						const insert = replacements.get(range);
-						if (insert === undefined) {
-							return {range};
-						}
-						return {
-							range: EditorSelection.range(range.from, range.from + insert.length),
-							changes: {from: range.from, to: range.to, insert},
-						};
-					}));
-				})();
-				return true;
-			}
-			return false;
+		run(): boolean {
+			return escapeWiki(cm);
 		},
 	},
 ]);