+++ src/statusBar.ts
@@ -3,9 +3,24 @@ import {nextDiagnostic, setDiagnosticsEffect} from '@codemirror/lint';
 import type {EditorView} from '@codemirror/view';
 import type {Extension, SelectionRange} from '@codemirror/state';
 import type {Diagnostic} from '@codemirror/lint';
+import type {CodeMirror6} from './codemirror';
 import type {LintSource} from './lintsource';
 
 declare type Severity = 'error' | 'warning';
+
+declare interface MenuItem {
+	name: string;
+	isActionable(this: void, cm: CodeMirror6): boolean;
+	getItems(this: void, cm: CodeMirror6): HTMLDivElement[];
+}
+
+export const menuRegistry: MenuItem[] = [];
+
+const optionAll = /* @__PURE__ */ (() => {
+	const ele = document.createElement('div');
+	ele.textContent = 'Fix all auto-fixable problems';
+	return ele;
+})();
 
 function getLintMarker(view: EditorView, severity: Severity): HTMLDivElement;
 function getLintMarker(view: EditorView, severity: 'fix', menu?: HTMLDivElement): HTMLDivElement;
@@ -50,16 +65,29 @@ const updateDiagnosticsCount = (diagnostics: readonly Diagnostic[], s: Severity,
 const hasFix = (diagnostic: Diagnostic): boolean | undefined =>
 	diagnostic.actions?.some(({name}) => name === 'fix' || name.startsWith('Fix:'));
 
+const isItemActionable = (cm: CodeMirror6, {name, isActionable}: MenuItem): boolean =>
+	cm.hasPreference(name) && isActionable(cm);
+
+const toggleClass = (classList: DOMTokenList, enabled: boolean): void => {
+	classList.toggle('cm-status-fix-enabled', enabled);
+	classList.toggle('cm-status-fix-disabled', !enabled);
+};
+
+const getDiagnostics = (all: readonly Diagnostic[], main: SelectionRange): Diagnostic[] =>
+	all.filter(({from, to}) => from <= main.to && to >= main.from);
+
 const updateDiagnosticMessage = (
-	view: EditorView,
+	cm: CodeMirror6,
 	allDiagnostics: readonly Diagnostic[],
 	main: SelectionRange,
 	msg: HTMLDivElement,
-	menu?: HTMLDivElement,
 ): void => {
-	const diagnostics = allDiagnostics.filter(({from, to}) => from <= main.to && to >= main.from),
-		diagnostic = diagnostics.find(({from, to}) => from <= main.head && to >= main.head) ?? diagnostics[0];
-	if (diagnostic) {
+	const diagnostics = getDiagnostics(allDiagnostics, main);
+	if (diagnostics.length === 0) {
+		msg.textContent = '';
+	} else {
+		const diagnostic = diagnostics.find(({from, to}) => from <= main.head && to >= main.head) ?? diagnostics[0]!,
+			view = cm.view!;
 		msg.textContent = diagnostic.message;
 		if (diagnostic.actions) {
 			msg.append(...diagnostic.actions.map(({name, apply}) => {
@@ -74,54 +102,68 @@ const updateDiagnosticMessage = (
 				return button;
 			}));
 		}
-	} else {
-		msg.textContent = '';
 	}
+};
+
+const updateMenu = (
+	cm: CodeMirror6,
+	allDiagnostics: readonly Diagnostic[],
+	main: SelectionRange,
+	classList: DOMTokenList,
+	menu?: HTMLDivElement,
+	fixer?: LintSource['fixer'],
+): void => {
 	if (menu) {
-		menu.replaceChildren(
-			...[
-				...new Set(
-					diagnostics.filter(hasFix)
-						.map(({message}) => / \(([^()]+)\)$/u.exec(message)?.[1])
-						.filter(Boolean) as string[],
-				),
-			].map(rule => {
+		const actionable = menuRegistry.filter(item => isItemActionable(cm, item)),
+			fixable = new Set(
+				fixer && getDiagnostics(allDiagnostics, main).filter(hasFix)
+					.map(({message}) => / \(([^()]+)\)$/u.exec(message)?.[1])
+					.filter(Boolean) as string[],
+			);
+		if (actionable.length === 0 && fixable.size === 0) {
+			toggleClass(classList, false);
+			return;
+		}
+		toggleClass(classList, true);
+		const actions = actionable.flatMap(({getItems}) => getItems(cm)),
+			quickfix = [...fixable].map(rule => {
 				const option = document.createElement('div');
 				option.textContent = `Fix all ${rule} problems`;
 				option.dataset['rule'] = rule;
 				return option;
-			}),
-			menu.lastChild!,
-		);
+			});
+		if (fixable.size > 0) {
+			quickfix.push(optionAll);
+		}
+		menu.replaceChildren(...actions, ...quickfix);
 	}
 };
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-export default (fixer: LintSource['fixer']): Extension => showPanel.of(view => {
+export default (cm: CodeMirror6, fixer: LintSource['fixer']): Extension => showPanel.of(view => {
 	let diagnostics: readonly Diagnostic[] = [],
 		menu: HTMLDivElement | undefined;
-	if (fixer) {
-		const optionAll = document.createElement('div');
-		optionAll.textContent = 'Fix all auto-fixable problems';
+	if (!view.state.readOnly && (fixer || menuRegistry.length > 0)) {
 		menu = document.createElement('div');
 		menu.className = 'cm-status-fix-menu';
 		menu.tabIndex = -1;
-		menu.append(optionAll);
-		menu.addEventListener('click', ({target}) => {
-			if (target === menu) {
-				return;
-			}
-			(async () => {
-				const {doc} = view.state,
-					output = await fixer(doc, (target as HTMLDivElement).dataset['rule']);
-				if (output !== doc.toString()) {
-					view.dispatch({
-						changes: {from: 0, to: doc.length, insert: output},
-					});
+		if (fixer) {
+			menu.addEventListener('click', ({target}) => {
+				if (target === menu) {
+					return;
 				}
-				view.focus();
-			})();
-		});
+				(async () => {
+					const {doc} = view.state,
+						output = await fixer(doc, (target as HTMLDivElement).dataset['rule']);
+					if (output !== doc.toString()) {
+						view.dispatch({
+							changes: {from: 0, to: doc.length, insert: output},
+						});
+					}
+					view.focus();
+				})();
+			});
+		}
 		menu.addEventListener('focusout', () => {
 			menu!.style.display = 'none';
 		});
@@ -133,7 +175,8 @@ export default (fixer: LintSource['fixer']): Extension => showPanel.of(view => {
 		position = document.createElement('div'),
 		error = getLintMarker(view, 'error'),
 		warning = getLintMarker(view, 'warning'),
-		fix = getLintMarker(view, 'fix', menu);
+		fix = getLintMarker(view, 'fix', menu),
+		{classList} = fix.firstChild as HTMLDivElement;
 	worker.className = 'cm-status-worker';
 	worker.append(error, warning, fix);
 	message.className = 'cm-status-message';
@@ -143,24 +186,22 @@ export default (fixer: LintSource['fixer']): Extension => showPanel.of(view => {
 	dom.append(worker, message, position);
 	return {
 		dom,
-		update({state: {selection: {main}, doc, readOnly}, transactions, docChanged, selectionSet}): void {
+		update({state: {selection: {main}, doc}, transactions, docChanged, selectionSet}): void {
 			for (const tr of transactions) {
 				for (const effect of tr.effects) {
 					if (effect.is(setDiagnosticsEffect)) {
 						diagnostics = effect.value;
-						const fixable = !readOnly && Boolean(fixer) && diagnostics.some(hasFix),
-							{classList} = fix.firstChild as HTMLDivElement;
-						classList.toggle('cm-status-fix-enabled', fixable);
-						classList.toggle('cm-status-fix-disabled', !fixable);
 						worker.classList.toggle('cm-status-worker-enabled', diagnostics.length > 0);
 						updateDiagnosticsCount(diagnostics, 'error', error);
 						updateDiagnosticsCount(diagnostics, 'warning', warning);
-						updateDiagnosticMessage(view, diagnostics, main, message, menu);
+						updateDiagnosticMessage(cm, diagnostics, main, message);
+						updateMenu(cm, diagnostics, main, classList, menu, fixer);
 					}
 				}
 			}
 			if (docChanged || selectionSet) {
-				updateDiagnosticMessage(view, diagnostics, main, message, menu);
+				updateDiagnosticMessage(cm, diagnostics, main, message);
+				updateMenu(cm, diagnostics, main, classList, menu, fixer);
 				const {number, from} = doc.lineAt(main.head);
 				position.textContent = `${number}:${main.head - from}`;
 				if (!main.empty) {